+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Changed
+- KYC recovery without filling out a form
+
 ## [1.6.0] - 2019-12-25
 #### Fixed
 - A bug with displaying atomic swap details